+++ lib/hcloud/resources/server_type.rb
@@ -39,7 +39,7 @@ module HCloud
 
     attribute :locations, :server_type_location, array: true, default: -> { [] }
 
-    attribute :deprecation, :deprecation
+    attribute :deprecation, :deprecation, deprecated: true
 
     attribute :deprecated, :boolean, deprecated: true