+++ lib/hcloud/entities/record.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module HCloud
+  class Record < Entity
+    attribute :value
+    attribute :comment
+  end
+end